backend/requirements.txt
Django==3.2.2
Pillow==8.2.0
psycopg2==2.8.6
djangorestframework==3.12.4
django-cors-headers==3.7.0